fix: finish release lint cleanup
